refactor(package) : 검색 화면 main/detail 패키지 분리
