+++ feature/search/src/main/java/com/wepli/search/detail/SearchDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.wepli.search.screen
+package com.wepli.search.detail
 
 import android.annotation.SuppressLint
 import android.widget.Toast
@@ -34,10 +34,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import appbar.WepliAppBar
-import com.wepli.search.state.SearchEffect
-import com.wepli.search.state.SearchIntent
-import com.wepli.search.state.SearchUiState
-import com.wepli.search.viewmodel.SearchViewModel
+import com.wepli.search.detail.state.SearchEffect
+import com.wepli.search.detail.state.SearchIntent
+import com.wepli.search.detail.state.SearchUiState
+import com.wepli.search.detail.viewmodel.SearchViewModel
 import com.wepli.shared.feature.mock.songMockData
 import com.wepli.uimodel.music.SongUiData
 import common.WepliSpacer
